commented weekCategoryList; blocked on progress
